// apps/web/src/API.ts
export type CreateBacklogItemInput = {
  id?: string | null,
  title: string,
  type: ItemType,
  rating: number,
  createdAt?: string | null,
  owner: string,
  status: ItemStatus,
  image: string,
};

export enum ItemType {
  GAME = "GAME",
  MANGA = "MANGA",
  ANIME = "ANIME",
}


export enum ItemStatus {
  COMPLETED = "COMPLETED",
  NOT_STARTED = "NOT_STARTED",
  PENDING = "PENDING",
  DROPPED = "DROPPED",
}
// ...
export type BacklogItem = {
  __typename: "BacklogItem",
  id: string,
  title: string,
  type: ItemType,
  rating: number,
  createdAt: string,
  owner: string,
  status: ItemStatus,
  image: string,
  updatedAt: string,
};